natro_macro.py
#!/usr/bin/env python3
"""
Natro Macro for macOS (https://github.com/NatroTeam/NatroMacro)
Copyright © Natro Team (https://github.com/NatroTeam)

This file is part of Natro Macro. Our source code will always be open and available.

Natro Macro is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation,
either version 3 of the License, or (at your option) any later version.

Natro Macro is distributed in the hope that it will be useful. This does not give you the right to steal sections from our code, distribute it under your own name, then slander the macro.

You should have received a copy of the GNU General Public License along with Natro Macro. If not, please redownload from an official source.
"""

import sys
import os
import time
import signal
import threading
import psutil
import subprocess
from pathlib import Path
import json
import logging

# GUI and automation libraries
import pyautogui
import cv2
import numpy as np
from PIL import Image, ImageGrab
import Quartz
import ApplicationServices
from AppKit import NSWorkspace

# Import custom modules
from lib.roblox import RobloxController
from lib.image_search import ImageSearch
from lib.duration_from_seconds import duration_from_seconds, hms_from_seconds
from lib.enum.enum_int import EnumInt
from lib.enum.enum_str import EnumStr
from lib.hyper_sleep import hyper_sleep, sleep_ms, sleep_us
from lib.json_utils import JSON
from lib.now_unix import now_unix, now_unix_ms
from lib.inventory_search import InventorySearch
from lib.menu_manager import MenuManager
from lib.walk_system import WalkSystem
from lib.data.memory_match_data import MemoryMatchData
from paths.path_handler import PathHandler
from patterns.pattern_handler import PatternHandler

# Configure pyautogui
pyautogui.FAILSAFE = True
pyautogui.PAUSE = 0.1

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('natro_macro.log'),
        logging.StreamHandler(sys.stdout)
    ]
)
logger = logging.getLogger(__name__)

class NatroMacro:
    def __init__(self):
        self.script_dir = Path(__file__).parent
        self.lib_dir = self.script_dir / "lib"
        self.assets_dir = self.script_dir / "nm_image_assets"

        # Global window coordinates (equivalent to AHK globals)
        self.window_x = 0
        self.window_y = 0
        self.window_width = 0
        self.window_height = 0

        # Heartbeat thread
        self.heartbeat_thread = None
        self.running = False

        # Initialize controllers
        self.roblox = RobloxController(self)
        self.image_search = ImageSearch(self)
        self.inventory_search = InventorySearch(self)
        self.menu_manager = MenuManager(self)
        self.walk_system = WalkSystem(self)
        self.memory_match_data = MemoryMatchData()
        self.path_handler = PathHandler(self)
        self.pattern_handler = PatternHandler(self)

        # Initialize
        self.setup()

    def setup(self):
        """Initialize the macro environment"""
        logger.info("Initializing Natro Macro for macOS...")

        # Set working directory
        os.chdir(self.script_dir)

        # Check for administrator privileges (macOS equivalent)
        if os.geteuid() != 0:
            logger.warning("Running without administrator privileges. Some features may not work.")

        # Close any existing instances
        self.close_existing_instances()

        # Start heartbeat
        self.start_heartbeat()

        # Set up signal handlers
        signal.signal(signal.SIGINT, self.signal_handler)
        signal.signal(signal.SIGTERM, self.signal_handler)

    def signal_handler(self, signum, frame):
        """Handle shutdown signals"""
        logger.info(f"Received signal {signum}, shutting down...")
        self.running = False
        self.cleanup()
        sys.exit(0)

    def close_existing_instances(self):
        """Close any existing Natro Macro processes"""
        current_pid = os.getpid()
        for proc in psutil.process_iter(['pid', 'name', 'cmdline']):
            try:
                if proc.info['pid'] != current_pid:
                    cmdline = proc.info.get('cmdline', [])
                    if any('natro_macro.py' in str(cmd) for cmd in cmdline):
                        logger.info(f"Terminating existing instance (PID: {proc.info['pid']})")
                        proc.terminate()
                        proc.wait(timeout=5)
            except (psutil.NoSuchProcess, psutil.AccessDenied):
                continue

    def start_heartbeat(self):
        """Start the heartbeat monitoring thread"""
        self.heartbeat_thread = threading.Thread(target=self.heartbeat_monitor, daemon=True)
        self.heartbeat_thread.start()
        logger.info("Heartbeat monitor started")

    def heartbeat_monitor(self):
        """Monitor the main script and restart if necessary"""
        while self.running:
            # Check if main script is still running
            # This is a simplified version - in the full implementation,
            # this would monitor various aspects of the macro
            time.sleep(30)

    def get_roblox_window(self):
        """Find and return Roblox window information"""
        return self.roblox.get_roblox_hwnd()

    def get_roblox_client_pos(self, hwnd=None):
        """Update global window coordinates for Roblox client"""
        return self.roblox.get_roblox_client_pos(hwnd)

    def image_search(self, needle_path, haystack_region=None, variation=0):
        """
        Search for an image within the screen or specified region
        Equivalent to AHK's ImageSearch
        """
        if haystack_region:
            x1, y1, x2, y2 = haystack_region
            return self.image_search.imagesearch_in_region(needle_path, x1, y1, x2, y2, variation)
        else:
            return self.image_search.imagesearch_on_screen(needle_path, variation)

    def click_at(self, x, y, clicks=1, interval=0.1):
        """Click at specified coordinates"""
        try:
            pyautogui.click(x, y, clicks=clicks, interval=interval)
            return True
        except Exception as e:
            logger.error(f"Click error: {e}")
            return False

    def send_keys(self, keys):
        """Send keystrokes"""
        try:
            pyautogui.typewrite(keys)
            return True
        except Exception as e:
            logger.error(f"Send keys error: {e}")
            return False

    def run(self):
        """Main macro loop"""
        self.running = True
        logger.info("Natro Macro started successfully")

        try:
            while self.running:
                # Main macro logic would go here
                # This is a skeleton - full implementation would include
                # all the automation logic from the original AHK script

                # Check if Roblox is still running
                if not self.get_roblox_window():
                    logger.warning("Roblox window not found, waiting...")
                    time.sleep(5)
                    continue

                # Example: Update window position
                self.get_roblox_client_pos()

                # Sleep to prevent excessive CPU usage
                time.sleep(0.1)

        except KeyboardInterrupt:
            logger.info("Macro stopped by user")
        except Exception as e:
            logger.error(f"Macro error: {e}")
        finally:
            self.cleanup()

    def cleanup(self):
        """Clean up resources"""
        self.running = False
        logger.info("Cleaning up...")

        # Stop heartbeat thread
        if self.heartbeat_thread and self.heartbeat_thread.is_alive():
            self.heartbeat_thread.join(timeout=5)

        logger.info("Natro Macro stopped")

def main():
    """Main entry point"""
    try:
        macro = NatroMacro()
        macro.run()
    except Exception as e:
        logger.critical(f"Critical error: {e}")
        sys.exit(1)

if __name__ == "__main__":
    main()
