requirements.txt
pyautogui==0.9.54
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
psutil==5.9.6
pyobjc-framework-Quartz==9.2
pyobjc-framework-ApplicationServices==9.2
typing-extensions>=4.0.0
